requirements.txt
absl-py==2.1.0
astor==0.8.1
astunparse==1.6.3
blis==0.4.1
boto==2.49.0
boto3==1.9.144
botocore==1.12.144
cachetools==4.1.1
catalogue==1.0.0
certifi==2019.3.9
chardet==3.0.4
click==7.1.2
configparser==5.0.0
cymem==2.0.3
docker-pycreds==0.4.0
docutils==0.14
en-core-web-sm==3.7.1
fasttext==0.9.2
filelock==3.0.12
Flask==1.1.2
flask-cors==4.0.0
gast==0.3.3
gensim==3.7.3
gitdb==4.0.5
GitPython==3.1.8
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.8
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.9.4
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.7
keras-ordered-neurons==0.7.0
Keras-Preprocessing==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
mkl==2024.0.0
murmurhash==1.0.2
nltk==3.4.1
numpy==1.16.3
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
packaging==20.4
pathtools==0.1.2
Pillow
plac==1.1.3
preshed==3.0.2
promise==2.3
protobuf==3.12.4
psutil==5.7.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.5.0
pyenchant==3.2.2
pyparsing==2.4.7
python-dateutil==2.8.0
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.2.0
PyYAML==5.1
regex==2020.7.14
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.2.0
sacremoses==0.0.43
scikit-learn==0.21.2
scipy==1.12.0
sent2vec==0.2.0
sentencepiece==0.2.0
sentry-sdk==0.17.8
shortuuid==1.0.1
six==1.15.0
sklearn==0.0
smart-open==1.8.3
smmap==3.0.4
spacy==3.7.4
srsly==1.0.2
subprocess32==3.5.4
tensorboard==1.9.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
termcolor==1.1.0
thinc==8.2.3
tokenizers==0.8.1rc2
torch==2.2.1
TorchCRF==1.1.0
torchvision==0.17.1
tqdm==4.28.1
transformers==4.38.1
typing-extensions==4.9.0
urllib3==1.24.3
wandb==0.10.2
wasabi==1.1.2
watchdog==0.10.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
